move capi_stub.go
