+++ stub/capi_stub.go
@@ -1,4 +1,4 @@
-package v1beta1
+package stub
 
 import (
 	corev1 "k8s.io/api/core/v1"
